+++ client/src/components/page-template/PageTemplate.js
@@ -40,6 +40,7 @@ import {
   FirewallClassic,
   IbmCloudInternetServices,
   Voicemail,
+  IbmCloudSecurityComplianceCenterWorkloadProtection,
 } from "@carbon/icons-react";
 import f5 from "../../images/f5.png";
 import { arraySplatIndex, contains, getObjectFromArray } from "lazy-z";
@@ -97,6 +98,8 @@ const navIcons = {
   FirewallClassic: FirewallClassic,
   IbmCloudInternetServices: IbmCloudInternetServices,
   Voicemail: Voicemail,
+  IbmCloudSecurityComplianceCenterWorkloadProtection:
+    IbmCloudSecurityComplianceCenterWorkloadProtection,
 };
 
 let pageOrder = [