+++ client/src/lib/forms/disable-save.js
@@ -11,13 +11,7 @@ const {
   isWholeNumber,
   nullOrEmptyStringFields,
 } = require("lazy-z");
-const {
-  invalidName,
-  invalidIpCommaList,
-  invalidIdentityProviderURI,
-  isValidUrl,
-  validSshKey,
-} = require("./invalid-callbacks");
+const { isValidUrl, validSshKey } = require("./invalid-callbacks");
 
 /**
  * check multiple fields against the same validating regex expression
@@ -197,6 +191,8 @@ function disableSave(field, stateData, componentProps, craig) {
     "buckets",
     "object_storage",
     "cos_keys",
+    "scc_v2",
+    "profile_attachments",
   ];
   let isPowerSshKey = field === "ssh_keys" && componentProps.arrayParentName;
   if (containsKeys(disableSaveFunctions, field)) {
@@ -250,6 +246,8 @@ function disableSave(field, stateData, componentProps, craig) {
         ? componentProps.craig.appid.keys
         : contains(["policies", "dynamic_policies"], field)
         ? componentProps.craig.access_groups[field]
+        : field === "profile_attachments"
+        ? componentProps.craig.scc_v2.profile_attachments
         : componentProps.craig[field]
     ).shouldDisableSave(stateData, componentProps);
   } else return false;