+++ client/src/lib/forms/duplicate-name.js
@@ -23,7 +23,12 @@ const { getAllSecrets } = require("./utils");
 function hasDuplicateName(field, stateData, componentProps, overrideField) {
   let allOtherNames = [];
   let stateField = overrideField || "name";
-  if (field === "cis") {
+  if (field === "profile_attachments") {
+    allOtherNames = splat(
+      componentProps.craig.store.json.scc_v2.profile_attachments,
+      "name"
+    );
+  } else if (field === "cis") {
     allOtherNames = splat(componentProps.craig.store.json.cis, "name");
   } else if (field === "cis_dns_record") {
     allOtherNames = nestedSplat(